requirements.txt
colorama==0.4.6
comtypes==1.4.8
customtkinter==5.2.2
darkdetect==0.8.0
docx2pdf==0.1.8
fire==0.7.0
fonttools==4.55.3
lxml==5.3.0
numpy==2.2.1
opencv-python-headless==4.10.0.84
packaging==24.2
pdf2docx==0.5.8
PyMuPDF==1.25.1
PyPDF2==3.0.1
python-docx==1.1.2
pywin32==308
termcolor==2.5.0
tqdm==4.67.1
typing_extensions==4.12.2
